chore: move tasks to be under pkg
