+++ manager/service_job_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/absmach/propeller/manager"
-	"github.com/absmach/propeller/task"
+	"github.com/absmach/propeller/pkg/task"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
