+++ manager/workflow.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/absmach/propeller/pkg/dag"
 	"github.com/absmach/propeller/pkg/storage"
-	"github.com/absmach/propeller/task"
+	"github.com/absmach/propeller/pkg/task"
 )
 
 type WorkflowCoordinator struct {
