+++ pkg/scheduler/roundrobin.go
@@ -2,7 +2,7 @@ package scheduler
 
 import (
 	"github.com/absmach/propeller/pkg/proplet"
-	"github.com/absmach/propeller/task"
+	"github.com/absmach/propeller/pkg/task"
 )
 
 type roundRobin struct {
