+++ pkg/scheduler/scheduler.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 
 	"github.com/absmach/propeller/pkg/proplet"
-	"github.com/absmach/propeller/task"
+	"github.com/absmach/propeller/pkg/task"
 )
 
 var (
