+++ pkg/storage/memory_adapter.go
@@ -6,7 +6,7 @@ import (
 
 	pkgerrors "github.com/absmach/propeller/pkg/errors"
 	"github.com/absmach/propeller/pkg/proplet"
-	"github.com/absmach/propeller/task"
+	"github.com/absmach/propeller/pkg/task"
 )
 
 type memoryTaskRepo struct {
